KUDU-1096 (part 3): Evict unreachable followers
This patch adds support for evicting followers that have not
successfully responded for some time.

The current implementation considers only controller errors -- typically
network errors -- except for RemoteErrors failures, and any other
response an indication of liveness. This is a pretty conservative
approach for now.

If a follower is unresponsive for longer than the configured timeout
(the current default is 5 minutes), it will be evicted.

We special-case the scenario where there are only 2 voters in the
config. In that case, we will not attempt to evict an unreachable
follower, in order to avoid dropping down to a single replica if the
leader is partitioned.

Change-Id: I66e77a8505f3cc59f83e945a6edf984c821a0896
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/7986
Tested-by: jenkins
Reviewed-by: Todd Lipcon <todd@cloudera.com>