KUDU-1125 workaround. Disable fsync by default on Master WAL
The master currently does one catalog table write for every reported
tablet. In a cluster with a lot of tablets/servers, this can cause
the master to effectively _never_ be able to handle reported tablets,
never successfully create tables, etc.

Until we reduce the number of redundant log writes in this code path,
let's revert the default back to no-fsync.

Change-Id: I6a513d6f7e0d4da70a50ed12d7ea896e260bbecf
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/8011
Tested-by: Jean-Daniel Cryans <jdcryans@cloudera.com>
Reviewed-by: Jean-Daniel Cryans <jdcryans@cloudera.com>