KUDU-872: Raft leader must not accept config change until it has committed an op
This is required for the Raft config change protocol to be correct.
See https://groups.google.com/forum/#!topic/raft-dev/t4xj6dJTP6E

In order to test this, I added an itest which causes followers to reject
updates while still allowing them to vote. Once the cluster is in this
state, we wait for the leader to be elected and then attempt to remove
one of the followers from the cluster. The implementation should reject
the ChangeConfig request.

Change-Id: Iaa9d2158c3022fef40aa926f3c95558d3a68fb3f
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/8030
Tested-by: jenkins
Reviewed-by: Todd Lipcon <todd@cloudera.com>