KUDU-995. Do not start an election if not a member of the config
There are currently corner cases when a node may be a non-participant in
a raft config and know about it. One such case is when a node is removed
from a config and then added back again immediately. If the node is not
deleted in the mean time, it may be sent the config change op containing
its removal and have to process some number of additional intervening
ops before encountering its own proceeding ADD_SERVER config change op.
If the node gets partitioned or the leader goes down before it sees its
own ADD_SERVER op, the node will currently trigger an election.

This patch essentially cancels that election process if the node is not
a participant in its local config. This is a bug fix (and a bit of a
hack) to avoid the crash that ensues (see linked JIRA for more info).

In the future, when we support NON_VOTER operation for nodes, we should
revisit this approach and disable failure detection while a node is a
NON_PARTICIPANT as well.

Change-Id: I4fef3f949fb2a19231fdc155378ee71c3810b846
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/8034
Tested-by: jenkins
Reviewed-by: Todd Lipcon <todd@cloudera.com>