consensus: Don't block on locks for votes
Voting is somewhat slow, because it fsyncs disk twice. If the election timeout
is set to the default on the YCSB cluster, I see a lot of election storms --
lots of vote requests come in, RPC handlers back up, Update calls get dropped,
more elections start, etc.

This patch has a potential way we can ameliorate election storms -- if we get a
vote request but another thread's already in the middle of either handling a
vote or a normal leader update, we can just deny the vote.

It's definitely _safe_, because you can always deny a vote. But maybe it's not
a good idea for liveness.

Another idea might be to do a quick pre-flight check of the vote request
without holding update_lock, and if we're going to deny the vote, don't bother
grabbing the lock. If it looks like the vote is good, we could grab the lock,
re-check all the pre-reqs, and then vote for real.

Change-Id: Ib3fc114089f7156eba3989d8599040c391b711be
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/8035
Tested-by: jenkins
Reviewed-by: Michael Percy <mpercy@cloudera.com>