KUDU-1122. Java client got "stuck" after some errors running YCSB
An issue in the Java client is that we permit using a KuduTable even if it's not fully created. We
do have protections in AsyncKuduClient to check agaist this, but if we're coming from KuduSession
then those operations are parked in a "operationsInLookup" structure. This gets really bad when not
using AUTO_FLUSH_SYNC because it means the client can send operations non-stop and even fill the
heap if the table takes a long time getting created.

This patch fixes one of those two issues, namely ensuring that the table is ready to be used before
sending operations to it. It does so by looping on IsCreateTableDone as long as it returns
false, or if the whole thing times out.

This patch adds a single test case that found a bug when opening a table that was already created
(something we weren't really testing for directly). It was also extensively tested against a slow
master (fsyncs on).

Change-Id: I19c8b1aefe426daa66b6f65e7d0dc48cbefaac3f
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/8070
Tested-by: jenkins
Reviewed-by: Dan Burkert <dan@cloudera.com>