java: improve javadoc formatting, add a few missing annotations
This adds audience annotations for some inner classes that were not previously
annotated. The audience annotations do not propagate from the outer class to
the inner, so even if the outer class is annotated public, we still have to
annotate the inner class as well.

This also improves some formatting on javadoc to use paragraph tags.

Lastly, I made some protected class members package-private so they don't show
up in the javadoc.

Change-Id: I6ba7098313e9d313a2caae216c42bebcc7a79caa
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/8141
Tested-by: jenkins
Reviewed-by: Jean-Daniel Cryans <jdcryans@cloudera.com>