KUDU-1149:[client] log warning on scans which may use pruning
Partition pruning is not yet implemented, so any scan over a non-PK partitioned
table will result in a full table scan. To eliminate surprises and manage
expectations we log warning when performing queries that are currently full
table scans which may be optimized in the future via partition pruning.

testing: I verified that the correct warnings were triggered during flex
partitioning tests of both the Java and C++ clients, and that warnings were not
triggered during tests on simple partitioned tables.

Change-Id: I2c4d9a82e15980e9e468cc5861691d4b27a6a31a
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/8201
Tested-by: jenkins
Reviewed-by: Todd Lipcon <todd@cloudera.com>