all_types-itest: attempt to fix flakiness of verification
This test seems to fail sometimes when it reads fewer rows than
expected. I'm guessing this is because it's reading from a follower,
and we don't properly manage safe-time to ensure that the read is
up to date.

This adds a ClusterVerfier check before counting the rows, such that
we wait until all replicas converge.

Change-Id: I97b5e0dcaca7d26143e42c9039f83032335a6ef5
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/8196
Tested-by: jenkins
Reviewed-by: David Alves <david.alves@cloudera.com>