Fix backoff behavior in client RetryFunc
This fixes an issue identified by Binglin a while back: in RetryFunc, we were
using an integer for the wait time, starting with 1ms, and then multipying by
5/4. With integer math, this was resulting in no backoff at all.

This commit cleans up the logic a bit and adds a unit test that the backoff is
appropriate. I left the unit test assertion sloppy to try to prevent flakiness.
The test fails without the bugfix since it retries a few thousand times,
whereas it tends to fall in the middle of the allowed range with the fix in
place.

Change-Id: If6b25dd1e6d798439e066d89998a4d0ad615ae88
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/8212
Tested-by: Todd Lipcon <todd@cloudera.com>
Reviewed-by: Michael Percy <mpercy@cloudera.com>