KUDU-1156. Compactions always preferred above flushes when MRS usage <64MB
This patch modifies how we handle FlushOps that are smaller than the threshold but have data.
We now assign them a perf improvement that's 'elapsed_ms / 1h', and if that's over 1.0 then that's
what it's set to. From looking at a few clusters, 1.0 seems to be a number that compactions rarely
reach.

This patch also refactors tablet_peer_mm_ops a bit so we can test it.

Change-Id: I054b43e9820bf30c64f34108fd223a1d42f55114
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/8213
Reviewed-by: Todd Lipcon <todd@cloudera.com>
Tested-by: jenkins