KUDU-1138. CreateTable should only assign replicas to live servers, should error if not enough
This changes CreateTable to look at the last heartbeat time of the registered
tablet servers, and not try to assign tablets to servers which haven't
heartbeated recently. I reused the same flag that Mike had previously added for
the threshold on heartbeat recency.

Additionally, I added a pre-check that there are enough live tablet servers
before we actually try to create the table. This is to address a common
problem where a user might try to create replicated tables on a single-node
test cluster, and give them back a nicer error message.

In order to fix a test failure, I also had to change
ExternalMiniCluster::Restart so that it waits for all of the restarted tablet
servers to register with the master before returning. Otherwise, attempts to
create a table just after a restart were likely to fail as not all of the
tablet servers had registered.

Change-Id: I3d7bc62c3424357a4404488f4194afbfbadac38c
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/8215
Tested-by: jenkins
Reviewed-by: Todd Lipcon <todd@cloudera.com>