KUDU-1080 - Add a way to keep a scanner alive in between calls to NextBatch()
This adds a new method, KuduScanner::KeepCurrentScannerAlive() that
performs a new rpc, ScannerKeepAlive(), on the tablet server and
resets the scanner's last access time, causing it not to expire for
an additional ttl.

This is useful for use cases that might take arbitrary time in between
calls to KuduScanner::NextBatch(), such as the one mentioned in KUDU-1080,
making sure that the scanner lives until the next call. On the other hand
this also requires pro-activity from the users part, meaning that simply
hung clients will not force the tablet server to keep the scanners (and
related resources) alive.

Change-Id: I1e28857389404fb810456b5684b98051ee762e64
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/8216
Tested-by: jenkins
Reviewed-by: Jean-Daniel Cryans <jdcryans@cloudera.com>
Reviewed-by: Todd Lipcon <todd@cloudera.com>