cfile: fix an incorrect NotFound result when seeking between two cfile blocks
This fixes a bug which was causing the YCSB servers to periodically log error
messages like:

E0919 19:58:46.582502 35229 cfile_set.cc:353] CFileSet indicated upper bound
  was within range, but key iterator could not seek. CFileSet upper_bound =
  user1400562636366295513, enc_key = user1386514829394916766\x20

This error started showing up more frequently once we switched the default
block size for the key column to 4KB. The issue turned out to be that we were
seeking to a key which fell between two CFile blocks. In this case, the index
would point to the lower block, but then when we tried to seek within that
block, we'd get a NotFound error, since the seek point fell after the last
element in the block. We incorrectly propagated this error back up the iterator
chain.

It's likely this could cause incorrect (or no) data to be returned from a scan,
or at the least could cause a performance issue as predicate pushdown would
fail.

This commit adds new assertions to cfile-test which, after writing a file with
10,000 strings, performs a seek exactly to each string as well as to a
non-existing row which falls between each consective pair of strings. This test
failed before the patch.

Additionally, I ran YCSB locally and verified that the errors no longer print.

Change-Id: I5e91feef4577df214b4bdfdd8cb52b649c90c63c
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/8280
Tested-by: jenkins
Reviewed-by: David Alves <david.alves@cloudera.com>