Improve error messages for timed out writes
This changes RPC to have a nicer error message on timeout which includes the
RPC call name and the remote server. Additionally, it changes the Batcher code
and RPC retrier to include better contextual information in the returned
Status.

Some of this information was available in LOG messages, but typically the
Status returned to the user had very little actionable information on timeouts,
making diagnosis tricky.

Change-Id: I27283d770f4ac012306dfe0abb5c556d692ca549
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/8282
Reviewed-by: David Alves <david.alves@cloudera.com>
Tested-by: jenkins