docs: add a top-level TOC, add no-jekyll flag
This adds a top-level 'toc' on the right-hand TOC nav which points to
all of the main docs pages.

Additionally, I added a --no-jekyll flag to the docs build script. If you're
using 'jekyll serve' within the docs directory, the 'jekyll build' done by the
script isn't necessary and just confuses jekyll.

Change-Id: I13b45c45317a26610463496011142ed63c8353fc
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/8284
Reviewed-by: Misty Stanley-Jones <mstanleyjones@cloudera.com>
Tested-by: jenkins