KUDU-1078. Return Status::Incomplete when reading ahead of log
Workaround for KUDU-1078. We should try to prevent this condition in all
cases if possible.

This patch changes LogCache to return Status::Incomplete when reading
operations that are in the future, vs Status::NotFound for operations
which have been GCed. This allows us to differentiate between these two
conditions. When we encounter Incomplete when formulating a request for
a peer, we currently log a DFATAL and return an error, which in release
mode causes the ConsensusPeer driver to try again later to get a request
for the remote.

Adds a (disabled) itest which triggers the following bug:

- a leader falls asleep for a bit
- another node elects itself and replicates some ops
- then the leader comes back and tries to replicate an op, and gets back an
  INVALID_TERM response. However, in that response it also sees a newer
  last_received which is higher than anything it personally sent
  - we call NotifyObserversOfTermChange(), but that's async, so we don't step
    down quite yet
- meanwhile, before the notification runs, we try to generate another
  request, which ends up reading an op "from the future", since it's starting
  from the peer's last_received

Todd and Mike both contributed code to this patch.

Change-Id: Ia60bed07205b71d8aac5dd4f0a2e47006a7986a7
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/8038
Reviewed-by: David Alves <david.alves@cloudera.com>
Tested-by: jenkins