BugFix: Fix KeepAlive() and related client test
This fixes a bug in KuduScanner::KeepAlive() where we would only
send keep alive requests after the first request and not after subsequent
requests.

While making sure that the test also covered this case I found a bug
in the test itself, as the scanner ttl was set on the ScannerManager long
before the flag was reset on the test, meaning that the ttl was actually
the default and not the one set on the test.

I confirmed that, after the test was fixed, it failed before the actual
client fix with "Scanner NotFound" and didn't fail afterwards.

Change-Id: Ib49d763bf99f0bd2c7a8f21036d490187f8acc3f
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/8294
Reviewed-by: Todd Lipcon <todd@cloudera.com>
Tested-by: jenkins