KUDU-948: don't crash if the client doesn't set a timeout
This fixes a bug where the client would crash if no timeout was set.
The fix was to use a default timeout when none is set.

After (lightly) digging it's not totally obvious why a previous attempt
to reproduce the bug in client-test.cc didn't do it. Still, removing the
timeout in all_types-itest definitely crashes 100% of the time, thus
this patch does that. The crash disappears with the fix.

This patch was contributed to by David and Todd.

Change-Id: Ibee32f423eb8ae6858977d2d39a5c0a834aa419c
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/7599
Tested-by: jenkins
Reviewed-by: Todd Lipcon <todd@cloudera.com>