master: don't consider a tablet created until it has a leader
This fixes the tablet report processing code so that we don't transition a
tablet into 'RUNNING' until we've seen a tablet report that includes an elected
leader. This fixes an issue seen in bolt80 where, if we created a table while
the tablet servers were under a lot of load, a majority of replicas might time
out, but the master would still consider the tablet created.

This resulted in a tablet with a permanent minority which could never elect a
leader. When a client tried to write to this tablet it would retry looking for
a leader and eventually time out.

The new unit test emulates this scenario by starting three servers, shutting
down two, and creating a table. We ensure that the table doesn't get created
until we restart the other servers.

The fixed behavior broke a bunch of unit tests which were previously relying
on disabling leader election. With election disabled, we now time out creating
tables. So, these tests now set a flag which causes the catalog manager to
revert to its old behavior, for use in unit tests.

I also increased the timeout for creating tables to 30 seconds instead of 10,
since I found that when creating a big table, we'd sometimes see timeouts in
the 10 second range, and there doesn't seem to be much harm to making it a bit
more lenient.

Change-Id: If809e89924a1a48db11331996abf4067d7986cb2
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/8345
Reviewed-by: David Alves <david.alves@cloudera.com>
Tested-by: Todd Lipcon <todd@cloudera.com>