Un-Clouderafy the community documentation
This patch takes a first stab at making our community documentation less Cloudera-centric. This
mostly affects the installation page since the easiest way to install Kudu is via CM. Instead,
we now link to Cloudera's own documentation for the steps that involve CM.

The hole punching was moved to troubleshooting in order to cleanup the installation page.

The quickstart page is now re-aligned to be more about quickstarting and using the VM to try out
stuff.

The configuration page is now missing a big chunk related to CM which will reside on the
Cloudera website. The link should be added once that page is available.

Change-Id: I07742677bd2e1ffcdd573a5f2e4099c09924246a
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/8387
Reviewed-by: Misty Stanley-Jones <mstanleyjones@cloudera.com>
Tested-by: Jean-Daniel Cryans <jdcryans@cloudera.com>