Add ASL 2.0 to LICENSE.txt
We should have a copy of the ASL 2.0 at the top of the LICENSE.txt file
in order to make it clear what license Kudu is distributed under.

Also add NOTICE.txt file, which is typical in open-source projects in
order to avoid having to put a copyright notice on every file in the
distribution.

Change-Id: Icf81502c9c9bcce767feb5556a37f10d57500d54
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/8472
Reviewed-by: Todd Lipcon <todd@cloudera.com>
Tested-by: jenkins