Add ASL 2.0 header comments to source files
These changes were automatically performed by a Perl script based on
some regular expressions. This doesn't cover all of the files that have
a Cloudera copyright notice, but it covers 95% of them.

I will follow up with another couple of patches to deal with the files
that my regular expressions couldn't easily handle.

Change-Id: I86863d9195449e5788ac8cb235ce2f3b720484a4
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/8473
Tested-by: Michael Percy <mpercy@cloudera.com>
Reviewed-by: Todd Lipcon <todd@cloudera.com>