KUDU-939: ksck should have a nice usage message
This commit adds a usage paragraph describing the common options
when using ksck, and changes the no arg behavior to print the
usage string and select flags.

Additionally, 'ksck_main.cc' is renamed to 'kudu-ksck.cc', in order to
get meaningful output from 'kudu-ksck --helpshort' (show just ksck
related flags). This feature of gflags is hard-coded to look for a .cc
file sharing the binary name to determine which flags to show! It would
be nice to have --help default to just the ksck flags and require
--helpfull to get all of the flags, but this isn't possible with gflags.

Change-Id: Ib09f0ae6467a39243c6178cbd6c9f0e75eb1c06d
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/8485
Tested-by: jenkins
Reviewed-by: Michael Percy <mpercy@cloudera.com>