Update master branch to 0.6.0-SNAPSHOT
Now that we created a branch for 0.5.0, we should bump our dev builds
to 0.6.0-SNAPSHOT so that we don't confuse the two.

Note that I didn't update the docs to say 0.6.0 yet -- I figure we can
continue to use 'master' to build docs for now so that less stuff is
shifting around.

Change-Id: I2171a7ae75fa54ef46e6630d624cc363a90133db
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/8496
Tested-by: jenkins
Reviewed-by: Michael Percy <mpercy@cloudera.com>