docs: only add configuration reference for the servers
The configuration reference on the command line tools was pretty messy,
as it ended up documenting tons of server-side flags that just happened
to get pulled in. For tools that we care about, we should probably manually
write docs, or add the ability to manually specify the set of flags we
care about documenting for each tool.

Change-Id: Id7bcef674972c4b564604a6a42591becc2e7abe9
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/8511
Reviewed-by: Misty Stanley-Jones <mstanleyjones@cloudera.com>
Tested-by: Todd Lipcon <todd@cloudera.com>