flags: force some separation between flags
I found the flags very difficult to read since there was little separation
between the table for one flag and the label for the next. This hack
adds an {nbsp} between each flag. This isn't recommended by the AsciiDoc
powers that be, but seems to work and was easier than trying to find a
CSS selector that would apply only where we wanted it.

Change-Id: I39c93685aec8a0764b9f1feb4d0d5ef6860d332d
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/8512
Reviewed-by: Misty Stanley-Jones <mstanleyjones@cloudera.com>
Tested-by: Todd Lipcon <todd@cloudera.com>