Replace Repository tab on site with Fork me on GitHub ribbon
The Repository tab leading to a different site is jarring and kind of
startles me, leading me to quickly hit the back button on my browser and
generally annoying me. Let's just use a Fork me on GitHub ribbon for now
and add links to the source code in the usual places: community,
download, or whatever.

I found this ribbon at https://ar.al/scribbles/fork-me-on-github-retina-ribbons/
and I think it looks pretty nice.

Change-Id: Ia0ca5873fc252c18cfb769a377538b17374ed237
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/8518
Tested-by: Todd Lipcon <todd@cloudera.com>
Reviewed-by: Todd Lipcon <todd@cloudera.com>