KUDU-1186. Work around NTP not synchronized initialization issue
The Master initialization code makes a "first run" decision at the very
beginning of startup based on whether it sees tablet data on disk and
then assumes that if the tablet data is there then the rest of the
metadata is there. One possible cause for a partially created master
tablet is a crash on startup due to an unsynchronized clock. Since this
is likely to be a common issue people run into when trying to run Kudu
for the first time, this works around the issue by doing that check
early.

However, we should still rework that startup logic to be more robust in
a separate patch.

Tested this manually and it seems to do the trick.

Change-Id: I29525101297db3b3bfbfd540ed5a3ed824ff62b0
Reviewed-on: http://gerrit.cloudera.org:8080/1023
Reviewed-by: Todd Lipcon <todd@cloudera.com>
Tested-by: Todd Lipcon <todd@cloudera.com>