KUDU-1193. Flush glog files after server start
This works around the glog buffering issue by flushing the glog files at
startup. This is helpful because users will know that the processes have
successfully started when they first start an idle master or tablet
server process. These processes aren't doing anything so they will
otherwise buffer most of the log output and it may appear that the
process has hung during startup. The startup messages are also useful
in that they print the RPC and web UI addresses and ports.

Tested this locally and it's a big improvement.

Change-Id: I422511b61d2763341afc88f4a1fa15bed1bbd016
Reviewed-on: http://gerrit.cloudera.org:8080/1024
Reviewed-by: Todd Lipcon <todd@cloudera.com>
Tested-by: Todd Lipcon <todd@cloudera.com>