Add a trigger_gerrit script
This can be used to trigger a job on the internal Jenkins which builds an
upstream gerrit, posting +1/-1 results.

We'll move to some upstream Jenkins infrastructure soon so that public
contributors can see build results, etc, but this is a stop-gap so we can
continue to test patches before committing them.

Change-Id: I1ca8a919dd94263b258b1d2f6b9807a03ede0bed
Reviewed-on: http://gerrit.cloudera.org:8080/1041
Reviewed-by: Mike Percy <mpercy@cloudera.com>
Tested-by: Mike Percy <mpercy@cloudera.com>