Remove Cloudera-specific build/test tools
We used those scripts and configuration files when checking out the code and building
the binaries was the best thing we had to deploy Kudu. This isn't something useful for
others since it relies on internal tools, and soon we won't even be using them since we'll
do everything via Cloudera Manager (for those who wonder).

Change-Id: Ida39337d67cb1601c0fd0b3a6a2fc75b62dfe18b
Reviewed-on: http://gerrit.cloudera.org:8080/1081
Reviewed-by: Todd Lipcon <todd@cloudera.com>
Tested-by: Todd Lipcon <todd@cloudera.com>