consensus: if leader election is disabled, block on the lock for votes
1b78a7960616230f5cf77e3512cb7f9f34c45be8 added an improvement where,
if consensus was busy handling a vote or update call, it would
vote 'no' on any election, assuming that the candidate would back-off
and try again. This seems to work well in clusters, but caused
flakiness of many tests where we manually trigger an election
and assume it must succeed.

This patch changes the behavior so we only reject vote requests due
to the busy lock in the case that leader election is enabled.

Without this patch, raft_consensus-itest's TestAddRemoveServer failed
280/1000 attempts on TSAN. With the patch, it never failed.

Change-Id: Ic66e8516a9124c728175ee5a919efe5164891004
Reviewed-on: http://gerrit.cloudera.org:8080/1206
Reviewed-by: David Ribeiro Alves <david.alves@cloudera.com>
Tested-by: Todd Lipcon <todd@cloudera.com>