[java client] Right-size the responses list in batch CB
We had a TODO and while running ITBLL today I noticed that things were awfully slow and the MR
tasks were burning multiple CPUs each. Profiling found that we were creating gigantic lists
that we didn't need. Right-sizing them improves throughput by 4-5x (not kidding).

The problem was that we were multiplying the buffer space by the number of batches, but turns out
that the sum of them all is usually the buffer space itself.

Change-Id: I39a3a7a3f03ca34576d9f864b7f8446ded24c462
Reviewed-on: http://gerrit.cloudera.org:8080/1220
Reviewed-by: David Ribeiro Alves <david.alves@cloudera.com>
Reviewed-by: Todd Lipcon <todd@cloudera.com>
Tested-by: Internal Jenkins