Use HybridClock error of 0 on OS X
This patches HybridClock to always use a clock error of 0 usec on OS X. This is
ncessary because OS X does not provide a way to retrieve the NTP clock error in
the kernel. This change is safe as long as Kudu is not used in a distributed
setting with OS X hosts.

Change-Id: I43a10affdc3f394a101b9e4021ffa65312340cf8
Reviewed-on: http://gerrit.cloudera.org:8080/1243
Reviewed-by: David Ribeiro Alves <david.alves@cloudera.com>
Tested-by: Internal Jenkins