OS X: size_t specializations for trace event metrics
OS X uses a typedef to define the size_t type as unsigned long, whereas Linux
uses a #DEFINE. The fallout is that on OS X size_t and unsigned long are
different types, and template instantiations distinguish between the two. In a
lot of places we track metrics of type size_t, but we only provide an
instantiation for unsigned long. The two options for making this compile are to
add a metric template instantiation for size_t, or cast every size_t metric to
be a uint64_t. I opted to go for the former, since it touches less code, and the
casts are not always straightforward because of function bindings.

Change-Id: I4019ac8dd4915b1e9a880d2bc037327e46b49d58
Reviewed-on: http://gerrit.cloudera.org:8080/1283
Tested-by: Internal Jenkins
Reviewed-by: Todd Lipcon <todd@cloudera.com>