Add Thread::CurrentThreadId method.
This commit adds a cross-platform method to kudu::Thread which retrieves the
current thread ID. The existing Thread::PlatformThreadId is renamed to
Thread::UniqueThreadId to signify that it does not necessarily return the true
system tid.

This commit also standardizes the tid type as int64_t. pid_t is not a
cross-platform solution for thread IDs, since only Linux has no distinction
between process and thread IDs. int64_t seems to be the most compatible
solution.

Change-Id: If0fe36687c93b4e78ae66c850824a95768d07c92
Reviewed-on: http://gerrit.cloudera.org:8080/1315
Tested-by: Internal Jenkins
Reviewed-by: Todd Lipcon <todd@cloudera.com>