Make web site navbar prettier
This change improves the styling of the site navbar. It adds a
blue background hover effect on the non-active tab links and removes the
on-hover background color from the active tab, which seemed like a bug.

This also adds a top-level div class "kudu-site" that makes it easier to
override the default bootstrap CSS rules and consolidates some CSS rules
that were specified separately but affected the same elements.

Tested on Firefox & Chrome on Linux and also in developer mode on Chrome
to do a quick check for mobile devices.

Change-Id: Ia43b3663b1525e651f42c2c117acc56f4d12f08b
Reviewed-on: http://gerrit.cloudera.org:8080/1275
Reviewed-by: Misty Stanley-Jones <misty@apache.org>
Reviewed-by: Todd Lipcon <todd@cloudera.com>
Tested-by: Todd Lipcon <todd@cloudera.com>