Fix clearing issues on community and FAQ pages
These pages don't clear their content, which results in a weird grey
border going through their headers, which is a result of the top border
of the footer element. Add some extra padding to the body columns of
these pages and clear the footer to get rid of these artifacts.

Change-Id: Ie3222d1f2ab579f7729ae5cd08fb62f06ffb5068
Reviewed-on: http://gerrit.cloudera.org:8080/1281
Reviewed-by: Mike Percy <mpercy@cloudera.com>
Tested-by: Mike Percy <mpercy@cloudera.com>