rpc: fix handling of short recv() while reading frame header
This fixes a bug where, if recv() returned a partial frame header, we would
still fall through to trying to read the framed data, despite not actually
having fully read the header. This didn't reproduce in existing tests, since in
most cases the network throughput isn't saturated enough that we end up
fragmenting a TCP packet in the middle of the frame. However, in a new high
performance cluster we were able to trigger this issue after about 10 minutes
of a multi-threaded Impala bulk insert.

In order to reproduce the bug in unit tests, this patch adds a hidden flag which
randomly reduces the requested number of bytes passed to the recv() call, such
that it often returns fewer than the call site wanted. This quickly reproduces
the issue before adding the fix.

Change-Id: Id10c2e3c9f7dc0f17a8a6effa0be15e047e3c5b5
Reviewed-on: http://gerrit.cloudera.org:8080/1282
Reviewed-by: Jean-Daniel Cryans
Tested-by: Internal Jenkins