KUDU-1213. fix error message when NTP is unsynchronized
This fixes the handling of error codes returned from ntp_gettime. Despite some
misinformation on the internet, it turns out that ntp_gettime only sets 'errno'
when it returns -1. For unsynchronized clocks, it returns TIME_ERROR but does
not set errno.

I also improved the error message when a leap second is inserted so that it
won't spew to the logs.

Change-Id: If6768f06ca70c61ab97a9840734e32d0b27a9694
Reviewed-on: http://gerrit.cloudera.org:8080/1293
Reviewed-by: David Ribeiro Alves <david.alves@cloudera.com>
Tested-by: Internal Jenkins