KUDU-1210 (2/2): FATAL at startup if required CPU flags are not present
During startup, the tablet server and master now check that the SSSE3 and
SSE4.2 instruction sets are available, and log a FATAL error if not. This
prevents crashing with a SIGILL later at runtime.

The client is also modified to return a Status::NotSupported() if the CPU
flags are not available.

Change-Id: I222dc551f870a0da434d9587ccdc690d63cbddac
Reviewed-on: http://gerrit.cloudera.org:8080/1327
Tested-by: Internal Jenkins
Reviewed-by: Mike Percy <mpercy@cloudera.com>