krpc: remove krpc_sasl_path configuration option
A user reported recently that, when using a version of cyrus-sasl
installed in a non-standard location, it couldn't find the SASL plugins.
This is because we overrided the default SASL search path with our
own SASL_CB_GETPATH callback.

This Kudu-local SASL path used to be necessary when we built against
a version of SASL in thirdparty/. However, 152ff259de7ea44c81e17c59fd4a7c5f41ba712e
long ago changed us to use the system-installed SASL.

So, this commit simply removes our custom SASL path.

Change-Id: I3d0964e6b8d0291ab464c49286ac16bba81bbd02
Reviewed-on: http://gerrit.cloudera.org:8080/1330
Tested-by: Todd Lipcon <todd@cloudera.com>
Reviewed-by: Mike Percy <mpercy@cloudera.com>