[java client] Document and enforce against Operation reuse
Operations (inserts, updates, deletes) should not be reused since we buffer
them and don't make copies. This wasn't really obvious in the javadoc and
one way this can really manifest itself is by OOME when adding so many
strings that they occupy the whole heap.

This patch also makes it so that binary and string columns can be
overwritten. This required changing the way we store those columns,
which is now using a fixed size list pre-populated with nulls. A
quick benchmark showed that it wasn't noticeably slower, and we don't
need to write position and length in the row alloc anymore when
adding a string.

The way we now prevent reusing Operations is by "freezing" the PartialRow
once the Op is applied.

Change-Id: I9482497422a2228b0aaf8a80b50e6d113ba2b615
Reviewed-on: http://gerrit.cloudera.org:8080/1335
Reviewed-by: Dan Burkert <dan@danburkert.com>
Tested-by: Internal Jenkins
Reviewed-by: Todd Lipcon <todd@cloudera.com>