KUDU-1233: Ignore committed config change requests in WAL
In tablet bootstrap, when there are missing COMMIT messages for
replicate ops in the middle of the log, all the following COMMIT
messages (in OpId sequence order) are also dropped. For example,
consider this log:

  REPLICATE 1.1 NO_OP
  COMMIT 1.1
  REPLICATE 1.2 WRITE_OP
  REPLICATE 1.3 CHANGE_CONFIG_OP
  COMMIT 1.3
  REPLICATE 1.4 CHANGE_CONFIG_OP
  COMMIT 1.4

Note that there is no commit for 1.2 (applying a huge batch may take
some time). This is legal because local application of operations, once
consensus-committed, is asynchronous and done in parallel with other
operations. In this case, at tablet bootstrap time, we will drop COMMIT
messages from 1.2 onward and re-prepare them, then re-apply them.

The result is that change config operations may appear to be uncommitted
in the WAL when in reality they have already been applied to the
ConsensusMetadata (cmeta) file on disk.

Additionally, the remote bootstrap procedure currently proceeds as
follows:

1. Read superblocks and blocks from disk, snapshot them
2. Read WAL from disk, snapshot the segments
3. Read ConsensusMetadata from disk, snapshot it

This ordering is another reason why the ConsensusMetadata may appear to
be "ahead" of the WAL during startup.

This patch adds checks when replicating config change operations such
that we do not apply those changes to the consensus metadata if we
detect that the consensus metadata is "ahead" of the WAL in those cases.

Included is a test that first delays application of writes, in order to
delay generation of the COMMIT message associated with them, then
asynchronously inserts a new record. Then it applies a series of
configuration change operations, ultimately triggering a remote
bootstrap. This reproduces the situation seen in the original bug report
and the test fails without the changes to raft_consensus_state.cc

Change-Id: I90cf80410d35d3e8bb84343a37617c042c015534
Reviewed-on: http://gerrit.cloudera.org:8080/1366
Tested-by: Internal Jenkins
Reviewed-by: David Ribeiro Alves <david.alves@cloudera.com>
Reviewed-by: Todd Lipcon <todd@cloudera.com>