tracing: fix a rare race which could cause trace-test to fail
This fixes an issue which caused trace-test to fail ~6% of the time on our EC2
boxes. I was able to reproduce the failure by running 'stress -c 12' on an
m3.2xlarge while looping the test.

The issue was a race in the following sequence:

App Thread T1                Flusher thread
------------------------------------------
Check tracing flag, see
  it enabled
                             Disable tracing
                             Start flush (generation=1)
                             See that T1 has no TLEB,
                               skip it
See that there is no TLEB
  buffer. Create a new one
  with generation 1

                             FinishFlush (generation = 1)
                             Increment generation to 2

Add trace event to TLEB with
  generation 1

                             Enable tracing (gen 2)
Generate trace events into
the old-generation TLEB
                             Disable tracing
                             Flush (gen 2)
                             Skip T1's buffer (wrong generation)

The issue is that, when generating trace events during the second session, we
continued to write them into a thread-local event buffer (TLEB) with the old
generation number. Then, when flushing the second session, we didn't include
those events.

The fix is fairly simple: before reusing an already-established TLEB, we check
that its generation number is current, and delete it if not. This mirrors
the way the code works in the upstream Chromium source.

After the fix, I wasn't able to repro the test failure.

While I was in the code, I also noticed an unused OptionalAutoLock variable,
which I removed.

Change-Id: Ided690437963704adf428771dfa4757e3a77e2e0
Reviewed-on: http://gerrit.cloudera.org:8080/1342
Tested-by: Internal Jenkins
Reviewed-by: Mike Percy <mpercy@cloudera.com>