rpc: include a better error message if a connection shuts down with pending data
We've recently seen an issue where we suspect that the client and server
somehow got "desynchronized", and the client was awaiting an invalid length
RPC response. If the client is waiting for too much data, the call will
eventually time out. In that case, this patch will produce a warning message
when we close the connection.

To test, I manually changed the server to send a response header which indicated
the length was 1 byte longer than the true length. This caused the client to
time out trying to read the extra byte. The resulting error message was:

W1028 11:48:18.732810 26275 connection.cc:135] Shutting down connection client
  Connection to 0.0.0.0:46067 with pending inbound data (12/13 bytes received,
  last active 500 ms ago, status=Service unavailable: reactor is shutting down
  (error 108))

Change-Id: I60a41b6a6fe3aa67b73587d1b806497cf47f74e7
Reviewed-on: http://gerrit.cloudera.org:8080/1345
Tested-by: Internal Jenkins
Reviewed-by: Mike Percy <mpercy@cloudera.com>